# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MIZ API
  version: 0.1.0
  description: The authoritative HTTP contract between the Rust API and MIZ clients.
servers:
  - url: /
security:
  - sessionCookie: []
paths:
...
  /api/v1/admin/auth/login:
    post:
      operationId: operatorLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorLoginRequest' }
      responses:
        '200':
          description: Operator signed in with MFA
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/admin/me:
    get:
      operationId: getCurrentOperator
      security:
        - operatorSessionCookie: []
      responses:
        '200':
          description: Current operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operator' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/admin/reports/{reportId}:
    get:
      operationId: getAdminReport
      security:
        - operatorSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report with immutable evidence
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminReport' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: reviewAdminReport
      security:
        - operatorSessionCookie: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewReportRequest' }
      responses:
        '200':
          description: Updated report review state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminReport' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/admin/auth/mfa/enroll:
    post:
      operationId: enrollOperatorMfa
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperatorEnrollmentRequest' }
      responses:
        '204': { description: MFA enrollment completed }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/admin/operators/{operatorId}/roles:
    put:
      operationId: replaceOperatorRoles
      security:
        - operatorSessionCookie: []
      parameters:
        - name: operatorId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/ObjectId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceOperatorRolesRequest' }
      responses:
        '204': { description: Operator roles replaced }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /api/v1/appeals:
    post:
      operationId: createModerationAppeal
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAppealRequest' }
      responses:
        '201':
          description: Appeal created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Appeal' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
...
  /api/v1/admin/audit-logs:
    get:
      operationId: readAuditLog
      security:
        - operatorSessionCookie: []
      parameters:
        - { name: targetType, in: query, required: true, schema: { type: string, enum: [user, post, report, operator, appeal] } }
        - { name: targetId, in: query, required: true, schema: { $ref: '#/components/schemas/ObjectId' } }
        - { name: from, in: query, required: true, schema: { $ref: '#/components/schemas/Timestamp' } }
        - { name: to, in: query, required: true, schema: { $ref: '#/components/schemas/Timestamp' } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 100 } }
      responses:
        '200':
          description: Time- and subject-bounded audit records
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEntry' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
  /api/v1/users/me/deletion-requests:
    post:
      operationId: requestAccountDeletion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestAccountDeletion' }
      responses:
        '201':
          description: Account deletion requested and all sessions revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountDeletionRequest' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/posts:
    post:
      operationId: createPost
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostRequest' }
      responses:
        '201':
          description: Created post
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/posts/{postId}/replies:
    get:
      operationId: listReplies
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Replies ordered oldest first
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    post:
      operationId: createReply
      parameters:
        - $ref: '#/components/parameters/PostId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostRequest' }
      responses:
        '201':
          description: Created reply
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Post' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/posts/{postId}/reports:
    post:
      operationId: createReport
      parameters:
        - $ref: '#/components/parameters/PostId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateReportRequest' }
      responses:
        '200':
          description: Existing unresolved report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '201':
          description: Created report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/reports/{reportId}:
    get:
      operationId: getReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report status visible to its reporter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    patch:
      operationId: updateReport
      parameters:
        - $ref: '#/components/parameters/ReportId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateReportRequest' }
      responses:
        '200':
          description: Updated unresolved report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '412': { $ref: '#/components/responses/Problem' }
        '428': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/users/{targetUserId}/block:
    post:
      operationId: blockUser
      parameters:
        - $ref: '#/components/parameters/TargetUserId'
      responses:
        '204': { description: User blocked and Follow relationships removed }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: unblockUser
      parameters:
        - $ref: '#/components/parameters/TargetUserId'
      responses:
        '204': { description: Block removed }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /api/v1/users/{targetUserId}/mute:
    post:
      operationId: muteUser
      parameters:
        - $ref: '#/components/parameters/TargetUserId'
      responses:
        '204': { description: User muted }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    delete:
      operationId: unmuteUser
      parameters:
        - $ref: '#/components/parameters/TargetUserId'
      responses:
        '204': { description: Mute removed }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/follow-requests/{relationshipId}/accept:
    post:
      operationId: acceptFollowRequest
      parameters:
        - $ref: '#/components/parameters/RelationshipId'
      responses:
        '200':
          description: Accepted follow relationship
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FollowRelationship' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
  /api/v1/timelines/home:
    get:
      operationId: getHomeTimeline
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Following timeline ordered newest first
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPage' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-miz_session
    operatorSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-miz_operator_session
  parameters:
    PostId:
      name: postId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
    ReportId:
      name: reportId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
    TargetUserId:
      name: targetUserId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
...
    RelationshipId:
      name: relationshipId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ObjectId' }
    PageLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 30 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 2048 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 255 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    ObjectId:
      type: string
      pattern: '^[0-9A-Za-z]{22}$'
      minLength: 22
      maxLength: 22
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 timestamp normalized to UTC.
    Handle:
      type: string
      pattern: '^[0-9A-Za-z](?!.*__)[0-9A-Za-z_]{1,22}[0-9A-Za-z]$'
...
    RequestAccountDeletion:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password: { type: string, minLength: 1, maxLength: 128, format: password }
...
    AccountDeletionRequest:
      type: object
      required: [id, status, requestedAt, restoreUntil]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        status: { type: string, enum: [pending, cancelled, purging, purged, restored] }
        requestedAt: { $ref: '#/components/schemas/Timestamp' }
        restoreUntil: { $ref: '#/components/schemas/Timestamp' }
    OperatorLoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      oneOf:
        - required: [totpCode]
        - required: [recoveryCode]
      properties:
        username: { type: string, minLength: 3, maxLength: 64 }
        password: { type: string, minLength: 12, maxLength: 128, format: password }
        totpCode: { type: string, pattern: '^[0-9]{6}$' }
        recoveryCode: { type: string, minLength: 1, maxLength: 128 }
    Operator:
      type: object
      required: [id, username, roles, recentMfa]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        username: { type: string }
        recentMfa: { type: boolean }
        roles:
          type: array
          minItems: 1
          items: { type: string, enum: [support, moderator, seniorModerator, administrator, auditor] }
...
    OperatorEnrollmentRequest:
      type: object
      additionalProperties: false
      required: [enrollmentToken, totpCode]
      properties:
        enrollmentToken: { type: string, minLength: 32, maxLength: 128 }
        totpCode: { type: string, pattern: '^[0-9]{6}$' }
...
    ReplaceOperatorRolesRequest:
      type: object
      additionalProperties: false
      required: [roles, reason]
      properties:
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, enum: [support, moderator, seniorModerator, administrator, auditor] }
        reason: { type: string, minLength: 1, maxLength: 8192 }
...
    CreateAppealRequest:
      type: object
      additionalProperties: false
      required: [username, password, actionId, explanation]
      properties:
        username: { $ref: '#/components/schemas/Handle' }
        password: { type: string, minLength: 1, maxLength: 128, format: password }
        actionId: { $ref: '#/components/schemas/ObjectId' }
        explanation: { type: string, minLength: 1, maxLength: 8192 }
...
    Appeal:
      type: object
      required: [id, actionId, status, version, createdAt, reviewedAt]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        actionId: { $ref: '#/components/schemas/ObjectId' }
        status: { type: string, enum: [pending, upheld, overturned] }
        version: { type: integer, format: int64, minimum: 1 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        reviewedAt:
          oneOf:
            - { $ref: '#/components/schemas/Timestamp' }
            - { type: 'null' }
...
    AuditEntry:
      type: object
      required: [id, actorOperatorId, eventType, targetType, targetId, reason, beforeState, afterState, requestId, reportId, createdAt]
      properties:
        id: { type: integer, format: int64 }
        actorOperatorId:
          oneOf:
            - { $ref: '#/components/schemas/ObjectId' }
            - { type: 'null' }
        eventType: { type: string }
        targetType: { type: string }
        targetId: { $ref: '#/components/schemas/ObjectId' }
        reason:
          oneOf:
            - { type: string }
            - { type: 'null' }
        beforeState:
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: 'null' }
        afterState:
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: 'null' }
        requestId:
          oneOf:
            - { $ref: '#/components/schemas/ObjectId' }
            - { type: 'null' }
        reportId:
          oneOf:
            - { $ref: '#/components/schemas/ObjectId' }
            - { type: 'null' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
...
    CreatePostRequest:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          description: Between 1 and 500 Unicode grapheme clusters after trimming; at most 8192 UTF-8 bytes.
    Post:
      type: object
      required: [id, authorId, replyToPostId, content, effectiveVisibility, state, version, createdAt, updatedAt, editedAt]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        authorId: { $ref: '#/components/schemas/ObjectId' }
        replyToPostId:
          oneOf:
            - { $ref: '#/components/schemas/ObjectId' }
            - { type: 'null' }
        content:
          oneOf:
            - type: string
              minLength: 1
              description: Between 1 and 500 Unicode grapheme clusters; at most 8192 UTF-8 bytes.
            - { type: 'null' }
        effectiveVisibility: { type: string, enum: [public, followers] }
        state: { type: string, enum: [published, tombstone] }
        version: { type: integer, format: int64, minimum: 1 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
        editedAt:
          oneOf:
            - { $ref: '#/components/schemas/Timestamp' }
            - { type: 'null' }
    ReportReason:
      type: string
      enum: [spam, harassment, hatefulContent, violence, sexualContent, illegalOrDangerousTrade, personalInformation, copyright, other]
    CreateReportRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: { $ref: '#/components/schemas/ReportReason' }
        explanation:
          type: string
          minLength: 1
          description: Required for other; at most 500 Unicode grapheme clusters and 8192 UTF-8 bytes.
    UpdateReportRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        reason: { $ref: '#/components/schemas/ReportReason' }
        explanation:
          type: string
          minLength: 1
          description: At most 500 Unicode grapheme clusters and 8192 UTF-8 bytes.
    Report:
      type: object
      required: [id, targetId, targetType, reason, explanation, status, version, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        targetId: { $ref: '#/components/schemas/ObjectId' }
        targetType: { type: string, enum: [post, reply] }
        reason: { $ref: '#/components/schemas/ReportReason' }
        explanation:
          oneOf:
            - { type: string }
            - { type: 'null' }
        status: { type: string, enum: [received, inReview, actioned, noAction] }
        version: { type: integer, format: int64, minimum: 1 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
    ReviewReportRequest:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { type: string, enum: [inReview, actioned, noAction] }
        reason: { type: string, minLength: 1, maxLength: 8192 }
        removeContent: { type: boolean, default: false }
    AdminReport:
      type: object
      required: [id, targetId, targetType, authorId, reason, explanation, status, version, createdAt, updatedAt, evidenceContent, targetVersion, targetCreatedAt]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        reporterId:
          oneOf:
            - { $ref: '#/components/schemas/ObjectId' }
            - { type: 'null' }
        targetId: { $ref: '#/components/schemas/ObjectId' }
        targetType: { type: string, enum: [post, reply] }
        authorId: { $ref: '#/components/schemas/ObjectId' }
        reason: { $ref: '#/components/schemas/ReportReason' }
        explanation:
          oneOf:
            - { type: string }
            - { type: 'null' }
        status: { type: string, enum: [received, inReview, actioned, noAction] }
        version: { type: integer, format: int64, minimum: 1 }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
        evidenceContent: { type: string }
        targetVersion: { type: integer, format: int64, minimum: 1 }
        targetCreatedAt: { $ref: '#/components/schemas/Timestamp' }
...
    FollowRelationship:
      type: object
      required: [id, followerId, followeeId, status, createdAt, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/ObjectId' }
        followerId: { $ref: '#/components/schemas/ObjectId' }
        followeeId: { $ref: '#/components/schemas/ObjectId' }
        status: { type: string, enum: [pending, accepted, rejected, cancelled, removed] }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
...
    PostPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Post' }
        nextCursor:
          oneOf:
            - { type: string }
            - { type: 'null' }
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, code]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        code:
          type: string
          enum: [auth_required, csrf_failed, forbidden, internal_error, resource_not_found, version_conflict, precondition_required, problem_validation_failed, handle_conflict, handle_change_too_soon, reauthentication_required, session_limit_reached, rate_limited, invalid_credentials, idempotency_required, idempotency_conflict, content_empty, content_too_long, content_too_large, edit_window_expired, cannot_follow_self, invalid_state_transition, parent_not_visible, invalid_cursor, cursor_expired, target_not_visible, cannot_block_self, cannot_mute_self, cannot_report_own_content, report_reason_required, report_already_exists, report_not_editable, invalid_report_transition, operator_auth_required, operator_mfa_required, operator_mfa_stale, operator_permission_denied, appeal_reviewer_conflict, deletion_already_pending, deletion_not_pending, restoration_window_expired, purge_in_progress, operator_username_unavailable, appeal_already_exists]
        requestId: { type: string }